Refactor BMZ, change prediction axes validation
