+++ src/careamics/bioimage/__init__.py
@@ -1,7 +1,7 @@
 """Provide utilities for exporting models to BioImage model zoo."""
 
 __all__ = [
-    "build_zip_model",
+    "save_bioimage_model",
     "import_bioimage_model",
     "get_default_model_specs",
     "PYTORCH_STATE_DICT",
@@ -9,7 +9,7 @@ __all__ = [
 
 from .io import (
     PYTORCH_STATE_DICT,
-    build_zip_model,
-    get_default_model_specs,
     import_bioimage_model,
+    save_bioimage_model,
 )
+from .rdf import get_default_model_specs
