+++ src/careamics/bioimage/io.py
@@ -2,126 +2,23 @@
 from pathlib import Path
 from typing import Union
 
+import numpy as np
 import torch
 from bioimageio.core import load_resource_description
 from bioimageio.core.build_spec import build_model
-from bioimageio.spec.model.raw_nodes import Model
 
-from careamics.config.config import (
+from ..config.config import (
     Configuration,
 )
 
 PYTORCH_STATE_DICT = "pytorch_state_dict"
 
 
-def _get_model_doc(name: str) -> str:
-    """
-    Return markdown documentation path for the provided model.
-
-    Parameters
-    ----------
-    name : str
-        Model's name.
-
-    Returns
-    -------
-    str
-        Path to the model's markdown documentation.
-
-    Raises
-    ------
-    FileNotFoundError
-        If the documentation file was not found.
-    """
-    doc = Path(__file__).parent.joinpath("docs").joinpath(f"{name}.md")
-    if doc.exists():
-        return str(doc.absolute())
-    else:
-        raise FileNotFoundError(f"Documentation for {name} was not found.")
-
-
-def get_default_model_specs(
-    name: str, mean: float, std: float, is_3D: bool = False
-) -> dict:
-    """
-    Return the default bioimage.io specs for the provided model's name.
-
-    Currently only supports `n2v` model.
-
-    Parameters
-    ----------
-    name : str
-        Algorithm's name.
-    mean : float
-        Mean of the dataset.
-    std : float
-        Std of the dataset.
-    is_3D : bool, optional
-        Whether the model is 3D or not, by default False.
-
-    Returns
-    -------
-    dict
-        Model specs compatible with bioimage.io export.
-    """
-    rdf = {
-        "name": "Noise2Void",
-        "description": "Self-supervised denoising.",
-        "license": "BSD-3-Clause",
-        "authors": [
-            {"name": "Alexander Krull"},
-            {"name": "Tim-Oliver Buchholz"},
-            {"name": "Florian Jug"},
-        ],
-        "cite": [
-            {
-                "doi": "10.48550/arXiv.1811.10980",
-                "text": 'A. Krull, T.-O. Buchholz and F. Jug, "Noise2Void - Learning '
-                'Denoising From Single Noisy Images," 2019 IEEE/CVF '
-                "Conference on Computer Vision and Pattern Recognition "
-                "(CVPR), 2019, pp. 2124-2132",
-            }
-        ],
-        # "input_axes": ["bcyx"], <- overriden in save_as_bioimage
-        "preprocessing": [  # for multiple inputs
-            [  # multiple processes per input
-                {
-                    "kwargs": {
-                        "axes": "zyx" if is_3D else "yx",
-                        "mean": [mean],
-                        "mode": "fixed",
-                        "std": [std],
-                    },
-                    "name": "zero_mean_unit_variance",
-                }
-            ]
-        ],
-        # "output_axes": ["bcyx"], <- overriden in save_as_bioimage
-        "postprocessing": [  # for multiple outputs
-            [  # multiple processes per input
-                {
-                    "kwargs": {
-                        "axes": "zyx" if is_3D else "yx",
-                        "gain": [std],
-                        "offset": [mean],
-                    },
-                    "name": "scale_linear",
-                }
-            ]
-        ],
-        "tags": ["unet", "denoising", "Noise2Void", "tensorflow", "napari"],
-    }
-
-    rdf["documentation"] = _get_model_doc(name)
-
-    return rdf
-
-
-def build_zip_model(
+def save_bioimage_model(
     path: Union[str, Path],
     config: Configuration,
     model_specs: dict,
-) -> Model:
+) -> None:
     """
     Build bioimage model zip file from model specification data.
 
@@ -133,11 +30,6 @@ def build_zip_model(
         Configuration object.
     model_specs : dict
         Model specification data.
-
-    Returns
-    -------
-    Model
-        Bioimage model object.
     """
     workdir = config.working_directory
 
@@ -161,6 +53,15 @@ def build_zip_model(
     config_path = workdir.joinpath("config.pth")
     torch.save(config.model_dump(), config_path)
 
+    # create cover
+    # TODO: this is a hack to avoid an export error in bioimage.io when the
+    # inputs/outputs have singleton dimensions.
+    # load .npy files
+    np.load(model_specs["test_inputs"][0]).squeeze()
+    np.load(model_specs["test_outputs"][0]).squeeze()
+
+    # create cover array
+
     # Create attachments
     attachments = [
         str(optim_path),
@@ -183,7 +84,7 @@ def build_zip_model(
         model_specs["tags"].append("2D")
 
     # build model zip
-    raw_model = build_model(
+    build_model(
         output_path=Path(path).absolute(),
         **model_specs,
     )
@@ -194,8 +95,6 @@ def build_zip_model(
     scheduler_path.unlink()
     grad_path.unlink()
     config_path.unlink()
-
-    return raw_model
 
 
 def import_bioimage_model(model_path: Union[str, Path]) -> Path:
