+++ src/careamics/utils/__init__.py
@@ -5,8 +5,8 @@ __all__ = [
     "denormalize",
     "normalize",
     "get_device",
-    "check_array_validity",
     "check_axes_validity",
+    "add_axes",
     "check_tiling_validity",
     "cwd",
     "compile_model",
@@ -18,4 +18,4 @@ from .context import cwd
 from .metrics import MetricTracker
 from .normalization import denormalize, normalize
 from .torch_utils import compile_model, get_device
-from .validators import check_array_validity, check_axes_validity, check_tiling_validity
+from .validators import add_axes, check_axes_validity, check_tiling_validity
