+++ tests/utils/test_validators.py
@@ -1,6 +1,7 @@
+import numpy as np
 import pytest
 
-from careamics.utils import check_axes_validity
+from careamics.utils import add_axes, check_axes_validity
 
 
 @pytest.mark.parametrize(
@@ -43,3 +44,22 @@ def test_are_axes_valid(axes, valid):
     else:
         with pytest.raises((ValueError, NotImplementedError)):
             check_axes_validity(axes)
+
+
+@pytest.mark.parametrize(
+    "axes, input_shape, output_shape",
+    [
+        ("YX", (8, 8), (1, 1, 8, 8)),
+        ("ZYX", (8, 8, 8), (1, 1, 8, 8, 8)),
+        ("SYX", (2, 8, 8), (2, 1, 8, 8)),
+        ("SZYX", (2, 8, 8, 8), (2, 1, 8, 8, 8)),
+        ("TZYX", (2, 8, 8, 8), (2, 1, 8, 8, 8)),
+    ],
+)
+def test_add_axes(axes, input_shape, output_shape):
+    """Test that axes are added correctly."""
+    input_array = np.zeros(input_shape)
+    output = add_axes(input_array, axes)
+
+    # check shape
+    assert output.shape == output_shape
